+++ tests/test_model/test_network.py
